🚑️ Hotfix:  리다이렉트 방식 변경
